requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
python-dotenv
